Provide test data and a grid cell
This commit updates the intersects method in Box to accept a location as well as a lat,long text string. This comports with how the method was used in the previous application but also lets us manually test it from the controller. I have also transplanted the test data and the methods that calculate the grid cell for each applicant or position location.