Resolve #59 Implement Rehire Sites
The rehire sites table is here integrated into the main application so that we now can return information on whether a returning applicant has actually applied from the main applicants table. @allthesignals hard work in the rehire-cbo-api branch has been pulled into to augment the data from the table of sites and applicants from 2017 along with a normalization task for duplicate sites.

Two new controller actions let the user interface call a list of folks who might be rehired from the rehire_sites controller, which now includes information on applicants, and then a first_timers scope and action for applicants lets us return only applicants that are not returned by the rehire_sites controller. In concert these two endpoints should provide the UI with all the data it needs to display both new and returning applicants and the joined data.