Resolve #63 Include Past Participant Data
This resolves the issue of including data about past participants by joining the applicant table with the rehire_sites table on the icims_id for the applicant. If the applicant is a rehire then there will be present the site_name attribute when the applicant is queried. We also provide the site_name in the serializer for the front-end.