Change iCIMS Update Process
This updates the process for updating iCIMS to instead of dynamically updating iCIMS as users make changes will update iCIMS in a single task, which is to be run at a single point in time. First it will update everyone that selected a site to the ‘candidate employment selection’ status, then it will move the folks who are selected to the ‘selected by site’ status and associate them with the position. Finally it will update all the applicants not chosen by a site back to the external portal status.