Resolve #114 Offers now account for open positions
We need to be able to generate multiple offers for a position based on whether the position has any openings. This update changes the code to now allow a position to have multiple offers. It then checks the preference based on whether the new potential match is preferred over any existing waiting offers.

This commit also stops running the matcher if we run out of applicants.