Resolve #120 Use icims for lottery pool
This updates the applicant lottery pool to simply pull the number of people that are in the lottery from icims data instead of calculating it locally. To do this we pull down everyone who is marked lottery activated, and then attempt to place the people for whom we have open positions for.