Use Graphhopper for Travel Time Calculations
This feature creates a job to use GraphHopper for Travel Time score calculations. It does so by increasing the number of concurrent Postgres connections to 30 via sidekiq.yml and database.yml to allow more jobs to execute at once. It also creates an ActiveJob that builds the travel time score between a position and job, and a rake task to create these jobs in import.rb. We also move away from using Sidekiq’s inline testing feature to using fake!. We do not need to use fake! for ActiveJob but it is a better practice to use it for sidekiq jobs if we’re going to test those.

We also had to modify applicant.rb to output strings that work with Graphhopper params.