Update Outgoing Messages Job and Endpoint
Moves the outgoing messages worker to an ActiveJob job to be consistent with the other jobs.

Adds appropriate tests for the worker.

Namespace the route to API in order to follow the pattern set by other endpoints.

Update routing and views to reflect the new endpoint locaiton and the fact that