Add tests for users applying to site
This adds the request/integration tests for the situation where a user applies to a job site or removes their application from a job site.

It also removes the unused applicants_position Factory and unused routes.

It makes requisitions dependent: destroy with positions because we do not need/want unassociated requisition records hanging around if we destroy the req.