Add Lottery API Endpoints
This commit adds controllers for the lottery related tasks: build preference scores, matches, and lottery activated statuses.

We update the import_applicant_job.rb to set a default of lottery_activated: true for imported applicants so that if a user goes to run a lottery we run it with all possible applicants. We also add documenation to the README for this.

We also update routes to namespace remaining items to API and adding the relevant new controller actions.