Update to two status columns
This updates the requisitions model to have two different statuses:

applicant status defines the relationship of a requisition in relation to an applicant. Either the applicant has submitted an application/requisition for a job or has withdrawn it.

partner status describes the relationship of the partner to the applicant. This supplants the previous “picks” table. Either the partner is interested in an applicant or is not interested.