Update Offers to Use “Offer Sent” Enum
This commit updates the offer model to use “offer sent” to enumerate the difference between a created offer (where the enum is “nil”) and an offer that has had an email sent to a user (“offer sent”). This will allow a subsequent commit to then check the Offers table for Offers where accepted is not “offer_sent” and fire off those emails.