Resolve #194 Send Offer Emails to Offerees
After we have generated a series of offers in direct selection and/or lottery with the status “accepted: nil” we email all the offered folks their offers. When we email them we update their status to accepted: “offer_sent”.

This commit also refactors the ActiveJob queue adapter to test for the entire test suite to DRY up the code.