Resolve #174 Add Endpoint for Lottery Match Status
This adds two endpoints. One endpoint shows whether there are jobs in the match_lottey queue. If it is in the queue then the end user should not be able to queue another match job. The second endpoint shows whether sidekiq currently has workers running. If sidekiq has a worker running then the user interface should not allow the running of a new lottery match either.