Validate Email Addresses of Applicants
This adds a feature that validates the email addresses of applicants by using the Mailgun email validation API and then setting a flag in the database to mark whether the email address is valid according to mailgun or not. This should let YEE staff triage and intervene on getting working email addresses for young people without them.